Modularize IoManager
