+++ WalletWasabi.Tests/StoreTests.cs
@@ -9,6 +9,7 @@ using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using WalletWasabi.Io;
 using WalletWasabi.Stores;
 using Xunit;
 using Xunit.Sdk;
@@ -356,7 +357,7 @@ namespace WalletWasabi.Tests
 
 			// Single thread file operations.
 
-			IoManager ioman1 = new IoManager(file1);
+			DigestableSafeMutexIoManager ioman1 = new DigestableSafeMutexIoManager(file1);
 
 			// Delete the file if Exist.
 
@@ -397,18 +398,18 @@ namespace WalletWasabi.Tests
 			// Check digest file, and write only differ logic.
 
 			// Write the same content, file should not be written.
-			var currentDate = File.GetLastWriteTimeUtc(ioman1.OriginalFilePath);
+			var currentDate = File.GetLastWriteTimeUtc(ioman1.FilePath);
 			await Task.Delay(500);
 			await ioman1.WriteAllLinesAsync(lines);
-			var noChangeDate = File.GetLastWriteTimeUtc(ioman1.OriginalFilePath);
+			var noChangeDate = File.GetLastWriteTimeUtc(ioman1.FilePath);
 			Assert.Equal(currentDate, noChangeDate);
 
 			// Write different content, file should be written.
-			currentDate = File.GetLastWriteTimeUtc(ioman1.OriginalFilePath);
+			currentDate = File.GetLastWriteTimeUtc(ioman1.FilePath);
 			await Task.Delay(500);
 			lines.Add("Lorem ipsum dolor sit amet, consectetur adipiscing elit.");
 			await ioman1.WriteAllLinesAsync(lines);
-			var newContentDate = File.GetLastWriteTimeUtc(ioman1.OriginalFilePath);
+			var newContentDate = File.GetLastWriteTimeUtc(ioman1.FilePath);
 			Assert.NotEqual(currentDate, newContentDate);
 
 			/* The next test is commented out because on mac and on linux File.Open does not lock the file
@@ -421,12 +422,12 @@ namespace WalletWasabi.Tests
 			 * https://github.com/dotnet/corefx/issues/5964
 			 */
 
-			//using (File.OpenWrite(ioman1.OriginalFilePath))
+			//using (File.OpenWrite(ioman1.FilePath))
 			//{
 			//	// Should be OK because the same data is written.
 			//	await ioman1.WriteAllLinesAsync(lines);
 			//}
-			//using (File.OpenWrite(ioman1.OriginalFilePath))
+			//using (File.OpenWrite(ioman1.FilePath))
 			//{
 			//	// Should fail because different data is written.
 			//	await Assert.ThrowsAsync<IOException>(async () => await ioman1.WriteAllLinesAsync(lines));
@@ -466,8 +467,8 @@ namespace WalletWasabi.Tests
 			// Simulate file write error and recovery logic.
 
 			// We have only *.new and *.old files.
-			File.Copy(ioman1.OriginalFilePath, ioman1.OldFilePath);
-			File.Move(ioman1.OriginalFilePath, ioman1.NewFilePath);
+			File.Copy(ioman1.FilePath, ioman1.OldFilePath);
+			File.Move(ioman1.FilePath, ioman1.NewFilePath);
 
 			// At this point there is now OriginalFile.
 
@@ -483,7 +484,7 @@ namespace WalletWasabi.Tests
 			// Check recovery mechanism.
 
 			Assert.True(
-				File.Exists(ioman1.OriginalFilePath) &&
+				File.Exists(ioman1.FilePath) &&
 				!File.Exists(ioman1.OldFilePath) &&
 				!File.Exists(ioman1.NewFilePath));
 
@@ -493,12 +494,12 @@ namespace WalletWasabi.Tests
 
 			// Check if directory is empty.
 
-			var fileCount = Directory.EnumerateFiles(Path.GetDirectoryName(ioman1.OriginalFilePath)).Count();
+			var fileCount = Directory.EnumerateFiles(Path.GetDirectoryName(ioman1.FilePath)).Count();
 			Assert.Equal(0, fileCount);
 
 			// Check Mutex usage on simultaneous file writes.
 
-			IoManager ioman2 = new IoManager(file2);
+			DigestableSafeMutexIoManager ioman2 = new DigestableSafeMutexIoManager(file2);
 
 			await Task.Run(async () =>
 			{
@@ -516,7 +517,7 @@ namespace WalletWasabi.Tests
 			});
 
 			// TryReplace test.
-			var dummyFilePath = $"{ioman1.OriginalFilePath}dummy";
+			var dummyFilePath = $"{ioman1.FilePath}dummy";
 			var dummyContent = new string[]
 			{
 				"banana",
@@ -542,7 +543,7 @@ namespace WalletWasabi.Tests
 		{
 			var file = Path.Combine(Global.Instance.DataDir, nameof(IoTestsAsync), $"file.dat");
 
-			IoManager ioman = new IoManager(file);
+			DigestableSafeMutexIoManager ioman = new DigestableSafeMutexIoManager(file);
 			ioman.DeleteMe();
 			await ioman.WriteAllLinesAsync(new string[0], dismissNullOrEmptyContent: false);
